requirements.txt
torch==1.4.0
flask~=1.1.2
facenet_pytorch
imutils~=0.5.3
dlib~=19.21.1
pip~=20.3.3
Jinja2~=2.11.2
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
numpy~=1.19.4
six~=1.15.0
setuptools~=51.1.1
torchvision~=0.5.0
Pillow~=8.0.1
networkx~=2.5
matplotlib~=3.3.3
scipy~=1.5.4
opencv-python~=4.4.0.46
PyYAML~=5.3.1
MarkupSafe~=1.1.1
Cython~=0.29.21
pandas~=1.1.5